requirements.txt

numpy==1.21.6
pandas==1.3.5
rdkit-pypi==2022.9.5
dash==2.14.1
dash-bio==1.0.2